requirements.txt
huggingface_hub
python-dotenv
streamlit
langchain
chromadb
sentence-transformers